Refactor websocket ops in backend
