+++ marimo/_server/sessions.py
@@ -37,7 +37,7 @@ from marimo import _loggers
 from marimo._ast import codegen
 from marimo._ast.app import App, _AppConfig
 from marimo._ast.cell import CellConfig
-from marimo._messaging.message_types import Alert, KernelReady, serialize
+from marimo._messaging.ops import Alert, KernelReady, serialize
 from marimo._output.formatters.formatters import register_formatters
 from marimo._runtime import requests, runtime
 from marimo._server.api.status import HTTPStatus
